apps/employee_directory/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "employee-directory-app"
version = "0.1.0"
description = "CDK app for Employee Directory (learning/pro-style)"
requires-python = ">=3.10"
dependencies = [
  "aws-cdk-lib>=2.235.1,<3.0.0",
  "constructs>=10.0.0,<11.0.0",
  "pydantic>=2.0,<3",
  "PyYAML>=6.0,<7",
]


[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["employee_directory*"]
